Fixing a couple mergeV/E issues.
 Checking for illegal hidden keys.

 Refactoring searchVertices to allow subclasses to override search criteria.

 https://issues.apache.org/jira/browse/TINKERPOP-2931
